Updating Chromeless template
Our new AEP docs on dev.socrata.com use the chromeless template. In fact, they
are the only ones on the site that use the chromeless template. When the build
tries to pull in this template, liquid gets cranky about a commented out line.
This removes that line so that we can fix the build and get the changes shipped
for AOIC.